Moved riot leagues tests to own file
